requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dnspython==2.5.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-PyMongo==2.3.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonify==0.5
MarkupSafe==2.1.5
passlib==1.7.4
PyJWT==2.8.0
pymongo==4.6.1
python-dotenv==1.0.1
Werkzeug==3.0.1